Add ReferenceFormatter service
This introduces a new ReferenceFormatter interface, to format a
Reference into a block of Wikitext, as well as one implementation for
the only style we’ll support for now, for Data Bridge. Other styles will
most likely be added as separate classes (though they may share some
code), not via a “style” parameter on the existing implementation.

The configuration input of the DataBridgeReferenceFormatter is a new
class, WellKnownReferenceProperties, encapsulating the property IDs
known from the configuration settings added in change Ibe2c53fa35. The
task of grouping/splitting a reference by those properties is extracted
into a separate class, ByCertainPropertyIdGrouper; since future
reference styles may not need all well-known properties to be grouped,
that class is independent from the WellKnownReferenceProperties and
receives the properties to split by as a plain array.

The DataBridgeReferenceFormatter gets a SnakFormatter injected and
documents that it *should* generate SnakFormatter::FORMAT_WIKI; however,
while it could also assert this in its constructor (via getFormat()),
I’ve decided against this. Wikitext is an untrusted input format, and so
accidentally injecting any other SnakFormatter into this formatter
should not be a vulnerability in the same way that, for example,
injecting a Wikitext formatter into a formatter expecting sanitized HTML
would be a vulnerability; therefore, I don’t think guarding against this
mistake is worth the cost of checking the format every time.

Bug: T247412
Change-Id: I89177a3d110d700a1dd1be7fd39e98390ac83a5e