bridge: depend on 'vue' ResourceLoader module from core
vue was introduced into core as ResourceLoader module in 4c221a1

We should depend on it instead of the (now redundant) wikibase equivalent
to maximize the chance of it being cached with the user already. This
gains relevance the more micro frontends depend on it.

Also simplifies bridge's externalization configuration as it does not
need to alias modules anymore.

Bug: T247519
Change-Id: I1cf7933756287dfe223cebc3e00dfed3964afab8