Do not assume wgKartographerLiveData is always set
Note that just a few lines below, in line 174, is a check if the property
exists. Which means when this code was written it was clear this property
is not always set. However, this check is in JavaScript. The PHP code
assumes the property is always there, and produces a notice if it is not.

This patch silences this notice. This is done under the assumption that
all following code is fine with an empty array.

Bug: T247458
Change-Id: Ic77b2c6b33afdcbd4cb23b59cae2736e1feb6b5f