bridge: restructure z-index usage
Usage of z-index is notoriously confusing. This change tries to make it
a bit easier to work with by defining the values of all z-indexes in a
single place.

Note: This still leaves the issue of new created stacking contexts that
still can only be solved by diligence and inspection of individual
components.

Change-Id: I9bfb752e0deb70851cd3aa1bbaa89c8c8e698cfb