bridge: Wire up License
The absolutely positioned <div> in the App component needs
`width: 100%;` to prevent it being too small if the text is not wide
enough.

This adds testing for reshowing the License to one of the existing
saving tests as adding another test just for that seems excessive
considering the cost of browser tests.

Bug: T246231
Bug: T246230
Change-Id: I56351571488c90ab2bf90ae203ebae740d34c4d6