Replace trivial MockBuilder usage in tests with createMock()
createMock() disables the constructor by default.

The code does the same before and after this patch, which raises the
question: why touch it then? My answer: It's shorter and quite a bit
easier to read, in my opinion.

A follow-up might inline some of the now trivial 1-line methods. I
did not wanted to do this in the same patch, to not make it to
complicated. However, I'm willing to update this patch if you think
it should be done.

Change-Id: I61f273065d66b0366134d82d9d0872abb30bb848