Add ReferenceFormatterFactory
This factory can create a DataBridgeReferenceFormatter, and possibly
other styles of reference formatters in the future.

Because the specifics of creating SnakFormatters differ between repo and
client, this class is client-specific. The alternative would have been
to somehow inject finished SnakFormatters into the factory, but then
using the factory would be more complicated and the factory itself would
do even less, so I decided against it.

The test isn’t great – it essentially tests that the factory runs
exactly the code that it runs – but at least better than nothing.

Bug: T247412
Change-Id: Ia4bd257cd76d4f3f19f1cc6d0ca414419f9e8793