Add ApiFormatReference
This adds a new API module to format a reference to the client,
initially only for use in Wikidata Bridge, but with parameters (style
and outputformat) to accommodate future expansion. The module is only
available if Data Bridge is enabled on the client.

In the second example, P2 is chosen as the example property, since it’s
most likely to exist on the repository: P1 was blacklisted for a long
time (until change I755075ddd1, commit 3ef32cb3cc), so P2 would be the
first assigned property ID on wikis set up before that change. It won’t
necessarily be a string property ID, so the formatted reference may
actually indicate an error (“value type does not match property type”),
but there’s not much we can do about that.

Bug: T247413
Change-Id: I93616015b8de37996f84a2bd399c3b237d2cd18b