bridge: send references to API for formatting and store result
The API that is used here only allows to format a single reference per
request. That limitation is encapsulated inside the concrete repository
implementation as it shouldn't be the concern of the action how the API
is implemented in detail.

The language used to format the reference is the individual page's "page
content language" ("the client wiki's content" according to the story) -
it "can be different for each page"[1] so it needs to be passed (not
only for explicitness) as the backend can not magically figure it out
for us.

I think it is a bit unfortunate we conceil the language (coming from
`wgPageContentLanguage`) in service construction (like for
`ApiEntityLabelRepository`) as it heavily impacts the behavior of the
application and, at least to me, is part of the interface (cf.
`ReferencesRenderingRepository`: "how should the reference be
formatted") - it should really be part of application state instead; but
that can be a refactoring should we decide to do it.

[1]: https://www.mediawiki.org/wiki/Manual:Language#Page_content_language

Bug: T247414
Change-Id: I72f92dc711a9ca0053ebe6c1e817228960bb35e4