Log on unconfigured well-known reference properties
Give the wiki administrator a gentle hint (“info” level) that they might
want to configure the well-known reference properties if they’ve been
left at their default level but are being used.

Note that this only logs when the properties are used to create a
ReferenceFormatter, not when they are initialized. In practice, this
should mean that the log messages only appear when the wbformatreference
API is used, but not on other requests.

Change-Id: I34ea48e3be475114f8cac73ff83252cd876ebcd5