wbterms: When reaching an edge case, try reading master with CONN_TRX_AUTOCOMMIT
The edge case of not being able to read from or write to master happens
due to REPEATABLE_READ locking in MySQL and can lead to infinite loop.
To fix this, you can get a CONN_TRX_AUTOCOMMIT connection to master to
make sure you read the value out of the transaction scope and
connections. Just trying to get a new connection will just re-use the
existing ones.

From documentation in LoadBalnacer::sanitizeConnectionFlags():
> Callers use CONN_TRX_AUTOCOMMIT to bypass REPEATABLE-READ staleness without
> resorting to row locks (e.g. FOR UPDATE) or to make small out-of-band commits
> during larger transactions. This is useful for avoiding lock contention.

Good luck being able to write a test for this.

Bug: T247553
Change-Id: I0649d715f8e7f6f1aa47cbb2dc113e73faddd42e