Set concepturi metadata in ApiEntitySearchHelper
This uses the concept URI from the wbsearchentities api result in
ApiEntitySearchHelper in the metadata array of the term search result
so that we do not have to retrieve it again.

This also factors out the addition of concepturi from SearchEntities and
into ConceptUriSearchHelper.
This isn't 100% needed, but seperates things in a nice way and is
similar to PropertyDataTypeSearchHelper.
The ConceptUriSearchHelper is added in the SearchEntities constructor
as it is a requirement by the search API module and should ALWAYS be
done, so putting it in any wiring seems evil.

Bug: T248455
Change-Id: If01d1bcb303e3f952cc195489615256adf475af3