bridge: don’t remove default wrap output class
If the HTML is wrapped in an element with class="mw-parser-output", we
get external link icons from the skin “for free” – and this makes sense
as well, since the HTML is indeed output from the parser, after all.

And with the whole HTML wrapped in an outer <div>, there’s no wrapping
<p> to strip anymore. (There’s no Parser function to remove <p>s buried
more deeply within the HTML.) This does introduce some additional
spacing, though, so we need to override that with a bit of extra CSS in
the bridge.

It would also be possible to support both “block HTML with output class”
and “unwrapped inline HTML” as separate output formats (similar to
Message::FORMAT_PARSE and Message::FORMAT_BLOCK_PARSE), but I figure we
don’t need that extra complexity at the moment. Maybe it’ll be useful to
add that once we stabilize this API and encourage external use.

Bug: T248479
Change-Id: I367aa036e5ec8dcd6d4a1957711bffecfb89a64a