bridge: introduce SAVED ApplicationStatus
This new status is entered after the save is finished (rather than, as
previously, returning to READY). It will be used to display a “thank
you” screen in that case and allow the user to edit references on repo.

This temporarily leaves the app in an awkward state when using the dev
entry point: after saving, the app still returns to the normal editing
view, and it’s even possible to change the edit decision, but attempting
to change the data value in the text input field causes an error because
it’s no longer legal to dispatch the setTargetValue action. This will be
resolved once we wire up the “thank you” screen, at which point the app
will never again return to the normal editing view. (I don’t think we
should allow setTargetValue when the status is SAVED.)

Bug: T247952
Change-Id: I01e6eb0de16bcbd037888b29b0b10d82de55a681