bridge: rudimentary “saved” status handling in App
This avoids the awkward App state mentioned in the previous commit
(change I01e6eb0de1): instead of letting the App return to the normal
editing view, which will produce an error when the user tries to
actually edit anything, show an extremely rudimentary “saved” screen,
which will later be replaced with the proper “thank you” component.

Note that this doesn’t change the App’s event wiring yet, and so outside
the dev entry point, this will just produce a (more or less) brief flash
of “saved” before the whole page reloads itself. (Once the “thank you”
component is wired up, the “saved” event will only be emitted after
clicking the button there, rather than as soon as the save is finished.)

The “updates store and interface with API response after successful
saving” e2e test is completely removed, because without showing the edit
interface after save, it is no longer possible for the e2e tests to
check whether the store was updated or not; this is true both for this
rudimentary solution and for the real “thank you” screen. It would be
possible to add a separate e2e test that a different screen is being
shown, but I don’t think that’s necessary; once the event wiring is
updated for the “thank you” screen, the existing e2e tests that check
whether the page is reloaded will have to interact with that screen
anyways, so I don’t think it will be necessary to test it separately.

Bug: T247952
Change-Id: If44ebbc23cf4ab010d333e42aace68a3d83dfd30