ApiFormatReference: add basic tests
Some basic tests calling wbformatreference for real. To avoid having to
set up entities and involved configuration to test against this is
restricted to asserting the results are roughly in line what we expect -
an actual "happy path" (without any warnings) is not achievable this
way.

Checking if the repo is enabled seeem counter-intuitive as this is a
client API but it is used to indirectly assert if the database is setup,
as the tests can only be performed if that is the case. This means
this will - in case of broken assertions - fail in wikibase-repo-docker
instead of wikibase-client-docker as one might expect.

https://www.mediawiki.org/wiki/Talk:MediaWiki_API_integration_tests
looks like it may be the future but still can be quite an adventure
instead of getting us results quickly - so it was disregarded.

Bug: T238661
Change-Id: I10bd0b10eb7b91827635c14dcaa6902c08a81a93