DumpEntities: Fix DB group default override
We can only access Wikibase lib's WikibaseSettings
after Wikibase has been loaded, so let's try to
do this as late as possible (in a hook handler).
Also bail out, if we are still too late (and log
a warning).

No tests added as this kind of bootstrap code
is very hard to properly test.

This can be manually tested by intercepting (or
printing) the value passed into LoadBalancer::__construct
as $params['defaultGroup'], when running the dumpJson
script (with and without an explicit --dbgroupdefault)

Bug: T248612
Change-Id: I4bd7b7c7452ee4cb944e537e0cf5555a669e299c