bridge: distinguish between “close” and “cancel”
“close” is the general button type; “cancel” is the resulting action
only when the user is in the general editing view. (After saving, this
name would not make sense). Change most occurrences of “cancel” to
“close” accordingly.

This is a purely internal change; in particular, the message for the
button is still “cancel” (or a translation thereof), not “close”.
MediaWiki core doesn’t seem to have a generic “close” message similar to
the “cancel” one we currently use; we may want to introduce a
bridge-specific “close” message instead.

Bug: T247954
Change-Id: Idd5de80bc9ec9355a5a455bfc63e4e312b238893