bridge: make WritingEntityRepository take Entity + Revision
Instead of accepting an EntityRevision containing the new Entity data
together with the old revision ID, WritingEntityRepository now accepts
the new Entity data and the old EntityRevision data + revision ID
separately. This will allow us to compare the two sets of entity data
and determine which bits actually need to be sent to the backend.

The base revision is an optional parameter, since logically, it should
be possible to save an entity without knowing the base revision (though
it’s not recommended, as it means the backend can’t patch conflicts).

Also fixes some typos in ApiWritingRepository.spec.ts – seemed like a
good opportunity, since that file needed changes anyways.

Bug: T248473
Change-Id: I9e51a65ebbce1f2a3c1a64cb9cc127282f744f4e