bridge: let only top-level actions influence flow
postEntityLoad itself is a "hook" to run whenever the entity changed -
it should not influence the application flow (applicationStatus). This
is done in the top-level actions (initBridge & saveBridge) which know
about it (INITIALIZING->READY & SAVING->SAVED respectively).

The transaction INITIALIZING->READY is actually not happening yet like
described because we rely on the default application state (which is
INITIALIZING) but I think it would be wholesome; cf.
I81466f88755cb67fcca1736e2a9b3c94636e0413

Change-Id: Ie75a5acea251fc1c39d03db55c5639f496ec51ad