bridge: implement ability to relaunch
The App gains the ability to relaunch itself when it receives a 'reload'
event from ErrorUnknown (though that component does not emit it yet).
When this event is received, a new relaunchBridge action is dispatched,
which performs a full reset of the store (via three new mutations) and
then dispatches initBridge again.

To distinguish between events that go all the way to init and ones, like
the new relaunch, that are app-internal, the previous Events enum is
split into two, initEvents and appEvents.

Based-On: I81466f88755cb67fcca1736e2a9b3c94636e0413
Bug: T249118
Change-Id: Ibe1291c553a3495a203706477c1af1ff0989797e
Co-Authored-By: Pablo Grass <pablo.grass@wikimedia.de>