bridge: track unknown errors
Add a trackUnknownError() method to BridgeTracker, to track the
occurrence of an unknown error with a certain type in the tracking
system; add a trackApplicationErrorsAsUnknown() action to send the types
of all errors to that tracker method; and instruct ErrorUnknown to
dispatch that action when it is mounted (i. e., when ErrorWrapper hasn’t
mounted any more specific error component, therefore the errors must be
unknown).

Graphite metrics in Wikimedia seem to follow two main naming
conventions, from what I could see with a quick glance through the
Graphite Browser: snake_case, and camelCase. Transforming to snake_case
is easier than transforming to camelCase, so let’s do that to map our
original errors (mostly UPPER_SNAKE_CASE) to a more conventional format
for Graphite.

Bug: T249120
Change-Id: Ie69f38f877330baddcf3a16bc06791899cd5b64a