docs & refactor, Improve docs on external json format
Reorganizes some of the content in the main json doc
This also spreads a few little factual updates regarding where
the json representation comes from and what processes it goes through.
Also the fact that all of this applies for Item and Properties but
extensions such as Lexeme and MediaInfo may differ.

This also lightly refactors a single method in ResultBuilder to
make it easier to follow and understand.

Bug: T241215
Change-Id: I6aeadf9875476d768db17bf5b52b3f0fae47de7f