FP: Introduce service factory just for our api based services
This allows us to not create these services in the wiring which
increases testability.
This also removes some duplicate service creation code that we had
already introduced into the fed prop entity type wiring overrides.

I haven't created any tests in this patch for the wiring and this is
a pure refactor, but we can now easily introduce tests as we
introduce new services to the fed prop realm.

Change-Id: Ia0295eed7458fb801c3d52590ac19f1d1c372881