bridge: make License sticking to body top more comprehensible
The License component, when shown, is supposed to stick to the top of
the body of our App, i.e. it should not scroll away along with the body
content which can overflow the limited vertical space which is available
to it.

To do so, License is positioned absolutely - relative to the body.

In order to be able to use the body as reference for absolutely
positioned elements, this patch changes how the body achieves its growth
to the maximum available space which so far made use of
absolute positioning (and it can not be both relative and absolute at the
same time; adding another element seemed detrimental). But I consider
this change to the body not only a solution to a concrete problem but as
actually providing value in itself.

The previous use of position:fixed, documented as positioning the
element relative to the viewport or - depending on conditions - relative
to some ancestors[0], might act as an explanation why we saw mixed behavior
in dev and prod. However, this is not conclusive - I went with what, I
hope, works instead of unraveling the mystery.

[0]: https://developer.mozilla.org/en-US/docs/Web/CSS/position#fixed

Change-Id: If53eb1c2565ccd644b28c24d0f43fbfb60cfd59a