bridge: add pageUrl to AppInformation and store
The original URL of the page is needed for error reporting.

The integration test for init is a bit awkward because we can’t easily
change the URL at test runtime without access to the jsdom instance
(which is managed by Jest); we can set the URL in the config, but the
test just has to hard-code the same value as the config.

Bug: T249497
Change-Id: I076e3dfe428e3ca2e773415b28e33f1e98b0449a