bridge: relaunch bridge from ErrorUnknown
Relaunch the bridge if the user clicks the relevant button inside
ErrorUnknown - as an escape hatch for them to e.g. overcome a poor
internet connection.

The event "reload" was renamed in the process (to "relaunch") to stop
confusing ourself (as this is not about reloading the page as will be
added in T244361). It should be noted however, that the places where
renaming happened intentionally do not use the event from our events
enum - they may be called the same but are a contract between App and
its subcomponents, while the event in the events enum is a contract
between App and its surrounding.

Messages were kept untouched as they received attention from translators
already.

Bug: T249124
Change-Id: I3229f9089198b5ed88a9136d2edb63b6d32ffa22