bridge: update WikimediaUI Base and OOUI
Perhaps the most notable change is a slight tweak to the Base10 color to
increase contrast when used with Accent50 (T248393), though I don’t
think this combination occurs in Bridge. Also, the font Nimbus Sans L
was removed from the sans font family (T245467).

Also, the icons are apparently minified now, which is nice.

Note that you may need to run `npm run prepare` again after this change,
in order to recreate the SCSS version of wikimedia-ui-base.

Bug: T234070
Bug: T246117
Bug: T248393
Change-Id: I2e00d1f515308a4060e3a24a3cea6b4de9333aef