entityLoaded: fire depending on relevant pages, not other modules
Making the firing of hooks a second class citizen (depening on something
else) results in a double conditional (both the firing as well as the
listening depends on something). Instead, the hook should be a first
class citizen (should fire when ever it is possible) and others can
react to it as desired.

This also resolves the problem that during qunit tests, when no entity
page is executed, entityLoaded was run because other modules depended on
it, but the necessary JS config was not present, which in turn resulted
in superfluous and ill-shaped AJAX requests.

wikibase.entityPage.entityLoaded is kept as a depenceny of
wikibase.ui.entityViewInit for practical reasons until we can be fairly
confident, that parser cache has picked up its addition to
FullEntityParserOutputGenerator for all entity pages. => T251007

Bug: T250305
Change-Id: Ib676c5820ac9c8baf00216fab794978bb96df626