TR: Handle oldStatement being null in hook handler
When the user creates a new statement, the oldStatement argument of the
wikibase.statement.saved hook is null. Adjust the declarations of the
hook handler functions to correctly reflect this, then update everything
that receives an oldStatement to handle null as well. (That handling
consists of simply ignoring the event.)

Bug: T250070
Change-Id: I623895952d79ac98d4fb4399fe8462b9b70bd2e5