FP: Implement IdHtmlLinkFormatter without existance / title check
This is very similar to LabelsProviderEntityIdHtmlLinkFormatter in
regular wikibase repo, however it takes different services for
formatting and also does not currently have special handeling for
deleted entities.

The intention will be to add special handeling for deleted entities in a
followup, however it is highly likley that this will also involve a
different service to that used by
LabelsProviderEntityIdHtmlLinkFormatter.

It may make sense to recombine these 2 formatters once we account for
deletion formatting here, altering
LabelsProviderEntityIdHtmlLinkFormatter to also use the narrower service
that we are likely to create.

If we do not re merge these to classes then it may make sense to extract
some of the shared logic between them into EntityIdLabelFormatter (or
somewhere else). Both classes currently share EntityIdLabelFormatter as
a base class.

Bug: T250789
Change-Id: I6e90c56ac3a316eefd666ba9c1517d5e8367bfea