bridge: ReportIssue: use entityTitle instead of entity id
The entity id may not be available if the loading of the entity failed
(e.g. if it does not even exist).

We conveniently have the original entityTitle, from the link the user
clicked, available in the store - using this instead to illustrate the
input we tried which led to the error (may actually be more to the
point).

The browser test assertions are very iffy but currently I don't see a
cleaner way - will look into it during T249124.

Bug: T241126
Change-Id: Ia764df3dd6cd3bb055c349f788fb8434b3c8e790