bridge: Add capability to provoke generic save error to browser tests
Loss of network connection is likely one of the most common errors
during saving and should be simulated well with these functions.

As these methods might be used in multiple specs, an extra utility file
related to network aspects has been created. In particular, the network
must be re-enabled after each test to ensure that tests are not
affecting each other.

Documentation:
https://webdriver.io/docs/api/chromium.html#setnetworkconditions

Bug: T249715
Change-Id: I09c9e383f3957e56476fc604058bad2033dc6f69