bridge: selenium: create and use shortcuts to open the app
Repeated patterns should be avoided, especially if the respective
abstractions already exist.

This renames the utility `openBridgeOnPage()` to `openAppOnPage()` as it
does not make any guarantees that a bridge (i.e. app in its most
optimistic state) is indeed the result.

Instead, tests should follow up calls to `openAppOnPage()` with
appropriate waitFor*s/assertions (or utilities that do that, if
suitable) to verify that they entered the branch they are expecting.

Change-Id: I4e36c0116247d29f8f4ffd5300cf9985e6f7b953