Remove bad assertions for keys in SimpleCacheWithBagOStuff
In https://gerrit.wikimedia.org/r/c/583199 this test started correctly
calling makeKey on the internal BagOStuff instance and thus respecting
the contract of the class.

However the assertions for the supposed valid cache keys were left in
SimpleCacheWithBagOStuff.

makeKey can return cache keys that dont match the more restrictive regex
provided in SimpleCacheWithBagOStuff and thus cause things that can land
in the cache to throw instead.

SimpleCacheWithBagOStuff should not really be concerned with the
structure of the cache keys and thus should not check the validity.

Bug: T245396
Change-Id: Ic8d5a66b5737532ed0591de599d9464882b3c931