entityViewInit: don't express entityLoaded module dep
entityViewInit depends on the hook system and some module firing the
entityLoaded hook.

The entityLoaded module does have dependencies which are not modeled in
ResourceLoader and consequently expressing it as a dependency can result
in a broken dependency chain (if the entityLoaded module's preconditions
are not satisfied).

The entityLoaded hook is fired by explicitly loading the entityLoaded
module in FullEntityParserOutputGenerator (at the "top level").

Bug: T251007
Change-Id: I00a5d095659c7e7d086744939ecba3939c168d57