bridge: reportIssueTemplateBody: use entityTitle
The entity id may not be available if the loading of the entity failed
(e.g. if it does not even exist).

We conveniently have the original entityTitle, from the link the user
clicked, available in the store - using this instead to illustrate the
input we tried which led to the error (may actually be more to the
point).

This was fixed for the error displayed to the user in f4f2cef.

Bug: T241126
Change-Id: I5266ca7f3f66d135c4d5a26b182b3af2a55fda32