Move ClientHooks to \Wikibase\Client + client/includes/
In change I997b81018c (commit 217fb13056), I wrote that having the hooks
class outside of includes/ seems to be a convention of sorts, but more
recently in change Ibe2ac0c4e9 (commit debd0fad8f) we decided that we
don’t want to follow that convention, and move the hooks class into the
regular namespace and directory instead, which also seems to be common.

One apparently unrelated Phan suppression is removed because Phan
complains that it’s no longer necessary, though I have no clue why it
would only start to complain in this change.

Change-Id: If206eaebb2f9356de8098fdbb3776ee5e7c3c883