Move Action classes to right namespace
They’re used in WikibaseLexeme, so add temporary class aliases.

Unfortunately, MediaWiki core’s AutoLoaderStructureTest demands that not
only the old class name, but also the new class name, be listed in
autoload.php, even though that shouldn’t be necessary because the
classes can be autoloaded via PSR-4, and build/generateAutoload.php
doesn’t write those autoload entries. So once more we temporarily have
an autoload.php file not generated by build/generateAutoload.php. (This
used to be a problem in the past, but then it was fixed for a while.
Apparently it’s back now.)

Bug: T172368
Change-Id: I75135d6d600637427a3de91ed1657337203626a6