bridge: split out ReadingEntityRevisionRepository interface
The ReadingEntityRepository interface previously mixed two tasks:
getting the latest revision, and getting a specific one. However, not
every implementation will be able to support both: the wbgetentities API
does not support getting the data for a specific revision.

Instead, extract the specific revision part into a new interface,
ReadingEntityRevisionRepository. SpecialPageReadingEntityRepository
implements this interface (instead of, as before, simply ignoring the
revision parameter), since it’s not too hard to implement; a handful of
tests are added for this as well.

This also removes the ability of the entity module to initialize itself
from a given entity ID and revision, since that would require to have a
ReadingEntityRevisionRepository in the services, and we currently just
have a ReadingEntityRepository. Removing the ability seems like the best
solution to me, since we don’t need this ability at the moment, nor are
we likely to need it in the future, in my opinion.

Change-Id: I15dd624bfcb9ad5d67d8bbc69a65cb09626e0cf8