bridge: more thorough test:distnodiff
Previously, test:distnodiff did not detect changes in CSS files, because
it only diffed two pairs of JS files. We can’t just indiscriminately
diff all of dist/ (because not all files in there are committed, so diff
would complain about missing files), but we can diff the JS files, the
CSS file, and also the img/ directory for the icons.

Note that this means that you may need to clean your dist/img/ directory
after changes to the icons (e. g. a new OOUI version), or diff will
complain about additional files that are in your dist/img/ but not in
/tmp/dist/img/.

Also, diff the files produced by build:app right after that build: that
should slightly speed up failing builds, since there’s no need to wait
for build:init in that case.

Change-Id: I8bf86dd21fe0a115c5c245b85d7395ceb06e1101