docs: WikiPageEntityRevisionLookup::getLatestRevisionId and redirects
This relates to the linked ticket, but the "issue" here as described
seems to mostly be solved.
However when this redirect handeling was added in
I631a616c21ca83568aad549c0a79c4c8b00e37fc as part of T196882 it made the
definition of the interface unclear.

Looking at other implementations of this method everything seems to now
be in a somewhat consistent state.

Bug: T112658
Bug: T196882
Change-Id: I9eb28301d377e380fe220278c970d013b95ff7aa