bridge: Fix error handling in WritingApiRepository
This resolves the bug that errors in WritingApiRepository were not
handled correctly. Specifically, the 'no-such-revisionid' error was
handled by the error handler for generic jQuery errors. The root cause
is that when creating the WritingApiRepository, the error handling done
by ForeignApi was not taken into account.

This patch uses ApiCore in the WritingRepository to make use of the proper
error handling implemented therein.

Change-Id: I4f31c863b915475012e5d2618657861f5741e535