bridge: send user name from init to app
Get the user name from the config in init, pass it through to the app,
and there store it as a boolean flag of whether a warning for anonymous
edits should be shown or not.

The dev entry point acts like an anonymous user because I assume this
will be the more interesting case, especially once we make it possible
to permanently dismiss this warning (though that’s not part of T246676).

Bug: T249829
Change-Id: Ib57f3a56c5936c6df616ae02c638c3a6b9fff1c0