bridge, TR: doc: track dist/ size over time
Add a build script which can collect the size of dist/ files over time
from the Git history, and build a website which presents this data. Add
two build scripts to package.json which run that script for Data Bridge
and Tainted References, respectively; the results should show up on
doc.wikimedia.org automatically, by virtue of being included in docs/.

This makes use of the github graphql API for which an access token is
required. I created T254319 to get a better name for the environment
variable carrying the token.

Bug: T253204
Change-Id: Ida695052548ae4c5b0f537f65c17139ea1467af9
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>
Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Co-Authored-By: Pablo Grass <pablo.grass@wikimedia.de>