Alter EntityUrlLookup to allow for context relevant URLs
This will be used by federated properties to link to full urls
when default wikibase entities will want to link to relative urls.

The Title based lookup now uses Title::getLinkURL which is not used
in the code in Wikibase that we plan on replacing with this service,
however they do the same thing and this avoids us reimplementing the
same logic in Wikibase.

The API based lookup for now contains the assumption that all entities
where and API lookup is needed should return the full URL.
This is fine for now and probably won't need to be altered for a year or
so or more.

Bug: T251241
Change-Id: I88095101d03e11ad2a88ba31daffbb8787407c2f