Only use CachingPrefetchingTermLookup when we have a cache.
CachingPrefetchingTermLookup and UncachedTermsPrefetcher use the cache
object to pass data between themselves.
There are situations where MediaWiki has no local server cache
availible. In these situation MediaWiki uses an EmptyBagOStuff as the
cache. This thus breaks the assumption that a cache is used that at
least caches data for the request.

Rather than using the caching lookup here, just use the raw lookup and
avoid all of the extra layers of complexity. This raw lookup already has
the in memory cache implemented as class arrays so that multiple lookups
of the same thing don't end up happening.

This is imposible to create a check for in the
CachingPrefetchingTermLookup constructor currently as we can't type hint
or assert not EmptyBagOStuff there due to the indirection of using
SimpleCache.

For now I have created some phpdoc stating the requirement.
If we end up ever needing to construct these objects anywhere else then
we may want to create a Factory object that will always do the
constructing, and thus always handeling the cache logic which is
currently in the wiring.

Bug: T247196
Change-Id: I200f00c76e880e449d40d8db7ad9471f723466b3