bridge: fix width and line-breaking of ReportIssue again
In change I1c5bcbeb17 (commit fc44853f2b), I removed the max-width of
the ReportIssue component, mistakenly believing it to be unnecessary
because the text looked fine on both “desktop” and “mobile” (the main
widths I tested were 500px and 499px, respectively). However, it turns
out that on sufficiently narrow screens (e. g. 250px wide), the text
will start to spill out of its container without the max-width. To
prevent this, I think we need some max-width, and it needs to take the
marginForCenterColumn() mixin into account (because max-width: 100%
alone will not respect those margins); therefore, we need a calc() on
desktop, to subtract those margins, and then have to override it on
mobile. This isn’t great, but I don’t see a better solution yet.

The second problem with the component pointed out by UX was the presence
of a horizontal scrollbar (just for the ReportIssue) component on these
narrow screens. I had originally added this scrollbar (overflow-x: auto)
because I noticed that on narrow screens (the same situation where,
without the max-width, the text would start to spill), the browser
refused to further break a domain name across lines, and thought that
adding a scrollbar was preferable in this case (assumed to be rare)
rather than cutting off the domain. However, what we really want
(according to UX) is to tell the browser to break at any position in
that line if necessary, even within a domain name, for instance (which
is usually considered to be a single unbreakable word); overflow-wrap:
break-word seems to do that well.

(Side note: the removal of the max-width actually rendered the
overflow-x: auto ineffective – if the text was allowed to spill out of
its container, scrolling it horizontally was never necessary, if I’m not
mistaken. However, that was not the *correct* way to get rid of the
scrollbar – this commit hopefully fixes it properly.)

Bug: T248087
Change-Id: Ibdaa0a7df44b9c86452e2d7b44d89928e8391a0d