bridge: wire up WarningAnonymousEdit in App
Make the App show the WarningAnonymousEdit if the warning should be
shown, and connect its “proceed” event to a new store action that unsets
the corresponding store flag. Also, update the AppHeader to not show the
save button if the warning is shown.

In the non-browser tests, the mw.config mock is changed to mock a
non-null user name by default, since it would be annoying to make all
e2e tests click through the warning first; only one App tests needs to
go through that.

In the browser tests, the openBridgeOnPage() helper also dismisses the
warning if necessary; tests that don’t use it for various reasons (e. g.
running some JS between opening the page and clicking the edit link)
have to call the “dismiss” helper explicitly. Dedicated tests are also
added, of course.

Bug: T249832
Change-Id: Ifb5e973a44733525459b9b4a4a023d24a7628523