bridge: selenium: assert loading is shown at first
We download a couple of KB and the loading bar only shows after at least
1sec. A throughput of 1000kb/s should give the loading bar plenty of time
to appear but at the same time complete loading before the custom
timeout.

The Loading (.wb-db-load) component's (unused) getter "int", probably a
typo of "init", was removed with its unused selector. The way the
selectors are organized inside ROOT_SWITCH should be reconsidered, as
.wb-db-bridge does not really belong in there given the name ROOT_SWITCH
(DataBridge is not directly controlled by the "root switch") but it
makes sense in practice (telling apart the stages the application can
assume).

This also gives IndeterminateProgressBar inside of Loading a more
descriptive selector ("visibility" does not lend itself to identify an
element).

Bug: T249124
Change-Id: Ia561fc03e64cf73a9ab0e3d71d1b938397090aca