Stop assuming that string DB names mean non local DBs
This assumption was scattered around the code base a bit, but most
notabily in the new term storage code and the property info table.

Using "wikidatawiki" as a dbname when on wikidata wiki should mean
that everything that needs to local db should continue to work.
This patch removes this assumption everywhere.

This patch also takes the time to factor the term store writer service
constructions out of WikibaseRepo and into their own Factory.
This was done so that the check for the local entity source containing
the expected entity types could be tested easily, rather than in the
larger mix of service wiring code.

As some of this logic has moved around a bit this patch also removes the
no longer needed tests regarding if the writers are using the local
entity source and also if the local entity source has access to the
correct entity type.
This is now done at service construction time.

Some dead code was also found regarding the db name in some of the newer
term storage classes and that was also removed in this change.

Bug: T252019
Change-Id: Ie8822ebd3661a21cbe475f75cae2615cc2b206c6